refactor(protocol): hoist kept service types out of legacy/ (R3 stage 1)
Pure relocation, zero behavior change. The kept architecture (PIR query
clients, Harmony attach, ARC proofs) references a small set of types that
currently live inside the legacy/ mount slated for deletion:

- legacy/scope.rs moves to src/scope.rs unchanged (ProviderId, ScopeId,
  BackendId, WorkloadId, AuthScheme, derive_provider_id). The kept
  attach.rs, proof.rs, all four SDK query clients, and wasm client all
  use these identity types.
- legacy/auth.rs contributes OperationStartV1, HintTransport,
  AUTH_FRAME_CLASS_V1, MAX_AUTH_PROOF_LEN, and
  OPERATION_START_DIGEST_DOMAIN to a new src/operation.rs; auth.rs
  re-exports them so every existing import path keeps working. The
  operation-service mapping unit test moves with the type.

lib.rs re-exports are untouched, so no consumer changes. After this,
everything left in legacy/ is referenced only by deleted-world payment
machinery and can die with it in the later R3 stages.

Verified: cargo check --workspace (minus cln-rpc-guard) clean;
cargo test -p pir-service-protocol (158 passed) and cargo test -p runtime
(17 binaries) all green.
