+++ crates/protocol/service/src/scope.rs
